contributors/ttwf_seattle/jaffe/samples/ttwf-js-metadata-cache-example.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate</title>
    <link rel="author" title="Your Name" href="mailto:youremail@address.com">
    <link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="The transform should be translate(100px,100px)"> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- The metadata_cache below was generated by testharnessreport.js and then
         manually cut/paste into this file.  This is only required when
         per-test metadata is used (see below).  To generate the metadata_cache,
         launch the test file in a browser.  A link will appear to the source
         code that can then be added to the <head>  
    -->
    <script id="metadata_cache">/*
    {
      "Test1": {},
      "Test2": {}
    }
    */</script>     
</head>
<body>
    <div id="myDiv"></div>
    <div id="log"></div>
    <script>
        // Set the transform 
          // NOTE: The line below uses a prefix (webkitTransform instead of transform).  
          // This would not be allowed in a real W3C Test.  
          // For the purpose of this example it has been included.
        document.getElementById("myDiv").style.webkitTransform = "translate(100px,100px)";
 
        // Verify that the transform was set as expected
          // Note: The line below uses a CSS prefix (-webkit-transform instead of transform)
          // This would not be allowed in a real W3C Test.
          // For the purpose of this example it has been included.
        test(function() {assert_equals(document.getElementById("myDiv").style.getPropertyCSSValue("-webkit-transform").cssText,
                        "translate(100px, 100px)",
                        "transform should be translate(100px,100px)")}, 
                        "Test1"),
                        {assert: 'Per-test metadata. First call to test function', author: 'Some Person'};    
                        
        test(function() {assert_equals(document.getElementById("myDiv").style.getPropertyCSSValue("-webkit-transform").cssText,
                        "translate(100px, 100px)",
                        "transform should be translate(100px,100px)")}, 
                        "Test2"),
                        {assert: 'Per-test metadata. Second call to test function', author: 'Other Person'};
    </script>    
</body>
</html>
